+++ apps/total-typescript/src/hooks/use-convertkit.tsx
@@ -41,7 +41,6 @@ export const ConvertkitProvider: React.FC<
       const params = new URLSearchParams(window.location.search)
       const ckSubscriberId = params.get(CK_SUBSCRIBER_KEY)
       if (!isEmpty(ckSubscriberId)) {
-        confirmSubscriptionToast()
         removeQueryParamsFromRouter(router, [CK_SUBSCRIBER_KEY])
       }
       try {
@@ -80,22 +79,4 @@ export const ConvertkitProvider: React.FC<
 
 export function useConvertkit() {
   return React.useContext(ConvertkitContext)
-}
-
-const confirmSubscriptionToast = () => {
-  return toast(
-    () => (
-      <div>
-        <strong>Confirm your subscription</strong>
-        <p>
-          Please check your inbox for an email that just got sent. Thanks and
-          enjoy!
-        </p>
-      </div>
-    ),
-    {
-      icon: '✉️',
-      duration: 6000,
-    },
-  )
 }
